requirements.txt
markupsafe==2.0.1
Jinja2==2.11.3
pyyaml==6.0
google-api-python-client
google-cloud-storage